+++ app/build.gradle
@@ -33,6 +33,9 @@ dependencies {
     implementation "androidx.constraintlayout:constraintlayout:1.1.3"
     implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
     implementation 'com.fasterxml.jackson.module:jackson-module-kotlin:2.9.8'
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.2.1"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.2.1"
+
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
